Refactors config_builder to new file
